Fix some pesky bugs
Added a select function menu to the insert function button
There was an issue that when selected to a foreign function and opening a new file, the editor would crash.
There was a bug where inserting a function with functions inside of it would just paste contents of the function instead of the object. This was fixed by removing the ability to place functions inside of functions.